quizai/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.5.0
djangorestframework==3.15.2
mysqlclient==2.2.5
PyJWT==2.9.0
sqlparse==0.5.1
